Client - Disconnect no Reconnect

+++ Dice_Client_v2/src/main/java/ups_sp/Utils/Const.java
@@ -28,8 +28,8 @@ public class Const {
 
 
     // Timeouts
-    public static final int DISCONNECTED_CONNECTION = 30000; // 10 seconds
-    public static final int NO_CONNECTION = 5000; // 5 seconds
+    public static final long DISCONNECTED_CONNECTION = 30000; // 10 seconds
+    public static final long NO_CONNECTION = 5000; // 5 seconds
 
 
 }
